feat: Implement head pose estimation and image preprocessing pipeline
- Enabled `facialTransformationMatrixes` in `FaceLandmarkerHelper` and added logic to extract the 4x4 matrix.
- Implemented `calculateHeadPose` and `calculateLipRotation` logic in `FaceLandmarkerHelper.Companion` for testability.
- Implemented robust `ImageUtils` with `applyBlur` (box blur) and `applyHistogramEqualization`.
- Added unit tests for new image processing utilities and head pose calculation logic using Robolectric.
- Optimized `MainActivity` to use a cached dummy bitmap for VSR placeholder, reducing garbage collection churn.
- Updated `docs/TODO.md` to reflect progress on Phase 3 (Computer Vision Pipeline).
- Ensured build stability by resolving AndroidX/Jetifier issues and verifying tests with Robolectric.

+++ app/src/main/java/com/HereLiesAz/Liperty/utils/ImageUtils.kt
@@ -22,9 +22,9 @@ object ImageUtils {
         return image.toBitmap()
     }
 
-    fun rotateBitmap(bitmap: Bitmap, rotationDegrees: Int): Bitmap {
+    fun rotateBitmap(bitmap: Bitmap, rotationDegrees: Float): Bitmap {
         val matrix = android.graphics.Matrix()
-        matrix.postRotate(rotationDegrees.toFloat())
+        matrix.postRotate(rotationDegrees)
         return Bitmap.createBitmap(bitmap, 0, 0, bitmap.width, bitmap.height, matrix, true)
     }
